+++ e2e/lib/r13n_e2e.dart
@@ -1,0 +1,4 @@
+/// A Very Good Project created by Very Good CLI.
+library r13n_e2e;
+
+export 'src/r13n_e2e.dart';
