+++ e2e/lib/src/r13n_e2e.dart
@@ -1,0 +1,7 @@
+/// {@template r13n_e2e}
+/// A Very Good Project created by Very Good CLI.
+/// {@endtemplate}
+class R13nE2e {
+  /// {@macro r13n_e2e}
+  const R13nE2e();
+}
