+++ e2e/test/src/r13n_e2e_test.dart
@@ -1,0 +1,11 @@
+// ignore_for_file: prefer_const_constructors
+import 'package:r13n_e2e/r13n_e2e.dart';
+import 'package:test/test.dart';
+
+void main() {
+  group('R13nE2e', () {
+    test('can be instantiated', () {
+      expect(R13nE2e(), isNotNull);
+    });
+  });
+}
