+++ pubspec.yaml
@@ -16,6 +16,4 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mason: 0.1.0-dev.46
-  path: ^1.8.2
   very_good_analysis: ^3.1.0
