--- tool/e2e/lib/src/e2e.dart
@@ -1,7 +1,0 @@
-/// {@template e2e}
-/// A Very Good Project created by Very Good CLI.
-/// {@endtemplate}
-class E2e {
-  /// {@macro e2e}
-  const E2e();
-}
